requirements.txt
asgiref==3.9.2
Django==5.2.6
django-crispy-forms==2.4
pillow==11.3.0
sqlparse==0.5.3
tzdata==2025.2
psycopg2-binary
python-decouple==3.8
django-allauth==0.57.0
requests==2.32.3
PyJWT==2.10.1
cryptography==44.0.0
